+++ packages/react-server/src/render/render.tsx
@@ -34,6 +34,8 @@ import {quiltDataMiddleware} from '../quilt-data';
 import {getLogger} from '../logger';
 import {ValueFromContext} from '../types';
 
+import {fallbackErrorMarkup} from './error';
+
 export {Context};
 export interface RenderFunction {
   (ctx: Context): React.ReactElement<any>;
@@ -49,6 +51,7 @@ type Options = Pick<
 > & {
   assetPrefix?: string;
   assetName?: string | ValueFromContext<string>;
+  renderError?: RenderFunction;
 };
 
 /**
@@ -58,7 +61,11 @@ type Options = Pick<
  */
 export function createRender(render: RenderFunction, options: Options = {}) {
   const manifestPath = getManifestPath(process.cwd());
-  const {assetPrefix, assetName: assetNameInput = 'main'} = options;
+  const {
+    assetPrefix,
+    assetName: assetNameInput = 'main',
+    renderError,
+  } = options;
 
   async function renderFunction(ctx: Context) {
     const assetName =
@@ -139,12 +146,28 @@ export function createRender(render: RenderFunction, options: Options = {}) {
       }`;
 
       logger.log(errorMessage);
-      ctx.status = StatusCode.InternalServerError;
 
       // eslint-disable-next-line no-process-env
       if (process.env.NODE_ENV === 'development') {
         ctx.body = errorMessage;
       } else {
+        if (renderError) {
+          const [styles, scripts] = await Promise.all([
+            assets.styles({name: 'error'}),
+            assets.scripts({name: 'error'}),
+          ]);
+
+          const response = stream(
+            <Html manager={htmlManager} styles={styles} scripts={scripts}>
+              {renderError(ctx)}
+            </Html>,
+          );
+
+          ctx.body = response;
+        } else {
+          ctx.body = fallbackErrorMarkup;
+          ctx.set(Header.ContentType, 'text/html');
+        }
         ctx.throw(StatusCode.InternalServerError, error);
       }
     }