+++ packages/react-server/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+
+- 🛑 Replace `isomorphic-fetch` with `cross-fetch` as peer dependency. Consumer project should install `cross-fetch` in their project or use `@shopify/sewing-kit >= v0.131.0` ([#1497](https://github.com/Shopify/quilt/pull/1497))
 
 ## [0.14.0] - 2020-06-06